api: update 232

+++ runelite-api/src/main/java/net/runelite/api/hooks/DrawCallbacks.java
@@ -107,7 +107,7 @@ public interface DrawCallbacks
 
 	void swapScene(Scene scene);
 
-	default boolean tileInFrustum(Scene scene, int pitchSin, int pitchCos, int yawSin, int yawCos, int cameraX, int cameraY, int cameraZ, int plane, int msx, int msy)
+	default boolean tileInFrustum(Scene scene, float pitchSin, float pitchCos, float yawSin, float yawCos, int cameraX, int cameraY, int cameraZ, int plane, int msx, int msy)
 	{
 		return true;
 	}
